App: every layout region is an expression of the canvas inputs
layout.rs keeps three pixel inputs (the canvas, the viewport, the cell) and a few knobs
(map scales and tiles, pad button and gaps, band inset); the column, maps, HUD lines, pad,
band rows, and the menu grid derive from them. The combat panels are text rows, the count
row sits above the front silhouettes' feet (FRONT_FEET_Y, a const from the camera), the
defeat box is centred by cells, the modal's offsets are named, the pad label is centred in
its button, the clock fits a given width. No number changes: every existing test passes as
it was, which is the proof of the derivation.

+++ crates/omnis-app/src/viewport.rs
@@ -138,8 +138,8 @@ fn redraw(
     let sky = DrawOp {
         paint: Paint::Fill {
             color: backdrop,
-            width: VIEWPORT_SIZE.0,
-            height: VIEWPORT_SIZE.1,
+            width: u32::from(VIEWPORT_SIZE.0),
+            height: u32::from(VIEWPORT_SIZE.1),
         },
         x: 0,
         y: 0,
